+++ src/app/game/ui/components/scene/player-panel/player-panel.component.html
@@ -7,7 +7,7 @@
   <hr>
   <div class="row">
     <div class="col-2">
-      <img [src]="'../../assets/unit/' + player.avatar" class="card-img-top" alt="hero">
+      <img [src]="'../../../../../../assets/unit/' + player.avatar" class="card-img-top" alt="hero">
     </div>
     <div class="col-4">
       <div class="align-middle">Attack power: {{ player.power }}</div>
