+++ src/app/game/ui/components/scene/player-panel/player-panel.component.scss
