+++ src/app/game/ui/components/scene/scene.component.html
