+++ src/app/game/ui/components/scene/scene.component.scss
@@ -3,7 +3,7 @@
 
   .castle {
     height: inherit;
-    background-image: url("../../assets/scene/castle_bridge.webp");
+    background-image: url("../../../../../assets/scene/castle_bridge.webp");
     background-size: contain;
     padding-top: 220px;
     padding-left: 50px;
