+++ src/app/game/ui/components/unit/unit.component.html
@@ -5,7 +5,7 @@
       <span class="card-title">{{ unit.unit.title }}</span>
     </div>
     <div class="text-center">
-      <img [src]="'../../assets/unit/' + unit.unit.avatar" class="card-img-top" alt="hero">
+      <img [src]="'../../../../../assets/unit/' + unit.unit.avatar" class="card-img-top" alt="hero">
     </div>
     <ul class="list-group list-group-flush mt-2">
       <li class="list-group-item">
