+++ src/app/game/ui/pipes/count.pipe.ts
