+++ src/app/game/ui/pipes/max.pipe.ts
