+++ src/app/game/ui/pipes/sum.pipe.ts
