--- src/app/model/state/state.ts
@@ -1,6 +1,0 @@
-import {Unit} from "../unit/unit";
-
-export interface State {
-  heroes: Unit[];
-  villains: Unit[];
-}
