--- src/app/types/types.ts
@@ -1,2 +1,0 @@
-export type Team = 'Heroes' | 'Villains';
-export type Timer = ReturnType<typeof setTimeout>;
